[refactor] move utils.py into utils folder, update main.py for msconsole rename
